Cleanup mock MCP server
