+++ tests/mcp-servers/wanaku-performance-test-mock-mcp/src/main/java/ai/wanaku/tests/mcp/server/MockMcp.java
@@ -3,11 +3,13 @@ package ai.wanaku.tests.mcp.server;
 import jakarta.enterprise.inject.Instance;
 import jakarta.inject.Inject;
 
+import io.quarkiverse.mcp.server.Resource;
+import io.quarkiverse.mcp.server.TextResourceContents;
 import io.quarkiverse.mcp.server.Tool;
 import io.quarkiverse.mcp.server.ToolArg;
 import ai.wanaku.core.forward.discovery.client.ForwardRegistrationManager;
 
-public class MockTool {
+public class MockMcp {
     @Inject
     Instance<ForwardRegistrationManager> registrationManager;
 
@@ -25,5 +27,18 @@ public class MockTool {
             }
         }
         return "1234567890";
+    }
+
+    @Resource(uri = "file:///mock/data", description = "A mock resource that returns static data for testing")
+    TextResourceContents mockData() {
+        int delay = config.delay();
+        if (delay > 0) {
+            try {
+                Thread.sleep(delay);
+            } catch (InterruptedException e) {
+                Thread.currentThread().interrupt();
+            }
+        }
+        return TextResourceContents.create("file:///mock/data", "mock-resource-data-0987654321");
     }
 }
